feat: unplot datastream
